refactor : 테스트에 반복적으로 사용하는 함수들을 testUtils/index.js로 옮김
