+++ __tests__/ApplicationTest.js
@@ -1,33 +1,6 @@
 import App from "../src/App.js";
-import { MissionUtils } from "@woowacourse/mission-utils";
-import { EOL as LINE_SEPARATOR } from "os";
-
-const mockQuestions = (inputs) => {
-  MissionUtils.Console.readLineAsync = jest.fn();
-
-  MissionUtils.Console.readLineAsync.mockImplementation(() => {
-    const input = inputs.shift();
-
-    return Promise.resolve(input);
-  });
-};
-
-const getLogSpy = () => {
-  const logSpy = jest.spyOn(MissionUtils.Console, "print");
-  logSpy.mockClear();
-
-  return logSpy;
-};
-
-const getOutput = (logSpy) => {
-  return [...logSpy.mock.calls].join(LINE_SEPARATOR);
-};
-
-const expectLogContains = (received, expectedLogs) => {
-  expectedLogs.forEach((log) => {
-    expect(received).toContain(log);
-  });
-};
+import { EOL as LINE_SEPARATOR } from 'os';
+import { mockQuestions, getLogSpy, getOutput, expectLogContains } from '../testUtils/index.js'
 
 describe("기능 테스트", () => {
   test("모든 타이틀 출력", async () => {
